[Refactor]#51 - product 관련 패키지 이동 및 정리
