+++ src/main/java/nunu/orderCount/domain/product/model/dto/request/RequestUpdateProductDto.java
@@ -1,6 +1,5 @@
 package nunu.orderCount.domain.product.model.dto.request;
 
-import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import javax.validation.constraints.NotNull;
 import lombok.AccessLevel;
@@ -8,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import nunu.orderCount.domain.member.model.MemberInfo;
+import nunu.orderCount.domain.product.model.ProductDtoInfo;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
